Splits migrationsv2 actions and unit tests into separate files (#101200)
* Splits migrationsv2 actions and unit tests into separate files

* Moves actions integration tests
