+++ src/core/server/saved_objects/migrationsv2/actions/wait_for_reindex_task.test.ts
@@ -1,0 +1,56 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { catchRetryableEsClientErrors } from './catch_retryable_es_client_errors';
+import { errors as EsErrors } from '@elastic/elasticsearch';
+jest.mock('./catch_retryable_es_client_errors');
+import { elasticsearchClientMock } from '../../../elasticsearch/client/mocks';
+import { waitForReindexTask } from './wait_for_reindex_task';
+import { setWriteBlock } from './set_write_block';
+
+describe('waitForReindexTask', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  // Create a mock client that rejects all methods with a 503 status code
+  // response.
+  const retryableError = new EsErrors.ResponseError(
+    elasticsearchClientMock.createApiResponse({
+      statusCode: 503,
+      body: { error: { type: 'es_type', reason: 'es_reason' } },
+    })
+  );
+  const client = elasticsearchClientMock.createInternalClient(
+    elasticsearchClientMock.createErrorTransportRequestPromise(retryableError)
+  );
+
+  const nonRetryableError = new Error('crash');
+  const clientWithNonRetryableError = elasticsearchClientMock.createInternalClient(
+    elasticsearchClientMock.createErrorTransportRequestPromise(nonRetryableError)
+  );
+
+  it('calls catchRetryableEsClientErrors when the promise rejects', async () => {
+    const task = waitForReindexTask({ client, taskId: 'my task id', timeout: '60s' });
+    try {
+      await task();
+    } catch (e) {
+      /** ignore */
+    }
+
+    expect(catchRetryableEsClientErrors).toHaveBeenCalledWith(retryableError);
+  });
+  it('re-throws non retry-able errors', async () => {
+    const task = setWriteBlock({
+      client: clientWithNonRetryableError,
+      index: 'my_index',
+    });
+    await task();
+    expect(catchRetryableEsClientErrors).toHaveBeenCalledWith(nonRetryableError);
+  });
+});
